Enhancement: pass action and convoId to notification
The create message and create conversation routes were modified to also
pass the action: 'newMessage' and the conversationId to the notification
system so that when the user is on the messages view, it does not
refresh the entire view. It instead only updates the conversation if
it's view with the new messages. If the user is viewing another
conversation, the recent messages are updated with the latest message.